Rename test data classes inline with other modules like control, prefabs, testcases, etc
Rename *TestDataRepository* to *TestDataExplorer* to have a consistent naming conventions across different modules
