--- src/Pixel.Automation.TestData.Repository.Views/TestDataRepositoryHostView.xaml.cs
@@ -1,15 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace Pixel.Automation.TestData.Repository.Views
-{
-    /// <summary>
-    /// Interaction logic for TestDataRepositoryView.xaml
-    /// </summary>
-    public partial class TestDataRepositoryHostView : UserControl
-    {
-        public TestDataRepositoryHostView()
-        {
-            InitializeComponent();
-        }     
-    }
-}
