+++ src/Pixel.Automation.TestDataExplorer.ViewModels/AssemblyInfo.cs
