+++ src/Pixel.Automation.TestDataExplorer.ViewModels/MockTestDataExplorerViewModel.cs
@@ -1,13 +1,13 @@
 ﻿using Caliburn.Micro;
 using System.Diagnostics.CodeAnalysis;
 
-namespace Pixel.Automation.TestData.Repository.ViewModels
+namespace Pixel.Automation.TestDataExplorer.ViewModels
 {
     [ExcludeFromCodeCoverage]
     /// <summary>
     /// Dummy screen to show in TestDataRepository host when there is no automation project open and active
     /// </summary>
-    public class MockTestDataRepositoryViewModel : Screen
+    public class MockTestDataExplorerViewModel : Screen
     {
     }
 }
