+++ src/Pixel.Automation.TestDataExplorer.ViewModels/MoveToGroupViewModel.cs
@@ -11,7 +11,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Pixel.Automation.TestData.Repository.ViewModels;
+namespace Pixel.Automation.TestDataExplorer.ViewModels;
 
 /// <summary>
 /// Allow users to pick a new group name for data source
