+++ src/Pixel.Automation.TestDataExplorer.ViewModels/Pixel.Automation.TestDataExplorer.ViewModels.csproj
