+++ src/Pixel.Automation.TestDataExplorer.ViewModels/TestDataSourceResult.cs
@@ -1,7 +1,7 @@
 ﻿using Pixel.Automation.Core.TestData;
 using System;
 
-namespace Pixel.Automation.TestData.Repository.ViewModels
+namespace Pixel.Automation.TestDataExplorer.ViewModels
 {
     internal record TestDataSourceResult(TestDataSource TestDataSource, Type DataSourceType);    
 }
