+++ src/Pixel.Automation.TestDataExplorer.Views/AssemblyInfo.cs
