+++ src/Pixel.Automation.TestDataExplorer.Views/MockTestDataExplorerView.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="Pixel.Automation.TestData.Repository.Views.MockTestDataRepositoryView"
+﻿<UserControl x:Class="Pixel.Automation.TestDataExplorer.Views.MockTestDataExplorerView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:Pixel.Automation.TestData.Repository.Views"
+             xmlns:local="clr-namespace:Pixel.Automation.TestDataExplorer.Views"
              xmlns:iconPacks="http://metro.mahapps.com/winfx/xaml/iconpacks"          
              xmlns:controls="clr-namespace:MahApps.Metro.Controls;assembly=MahApps.Metro"  
              mc:Ignorable="d"
