+++ src/Pixel.Automation.TestDataExplorer.Views/MockTestDataExplorerView.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace Pixel.Automation.TestDataExplorer.Views
+{
+    /// <summary>
+    /// Interaction logic for MockTestDataExplorerView.xaml
+    /// </summary>
+    public partial class MockTestDataExplorerView : UserControl
+    {
+        public MockTestDataExplorerView()
+        {
+            InitializeComponent();
+        }
+    }
+}
