+++ src/Pixel.Automation.TestDataExplorer.Views/MoveToGroupView.xaml
@@ -1,4 +1,4 @@
-﻿<Controls:MetroWindow x:Class="Pixel.Automation.TestData.Repository.Views.MoveToGroupView"
+﻿<Controls:MetroWindow x:Class="Pixel.Automation.TestDataExplorer.Views.MoveToGroupView"
                       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                       xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
                       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
