+++ src/Pixel.Automation.TestDataExplorer.Views/NewDataSourceGroupView.xaml
@@ -1,4 +1,4 @@
-﻿<Controls:MetroWindow x:Class="Pixel.Automation.TestData.Repository.Views.NewDataSourceGroupView"
+﻿<Controls:MetroWindow x:Class="Pixel.Automation.TestDataExplorer.Views.NewDataSourceGroupView"
                       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                       xmlns:Controls="clr-namespace:MahApps.Metro.Controls;assembly=MahApps.Metro"
