+++ src/Pixel.Automation.TestDataExplorer.Views/Pixel.Automation.TestDataExplorer.Views.csproj
