+++ src/Pixel.Automation.TestDataExplorer.Views/RenameDataSourceGroupView.xaml
@@ -1,4 +1,4 @@
-﻿<Controls:MetroWindow x:Class="Pixel.Automation.TestData.Repository.Views.RenameDataSourceGroupView"
+﻿<Controls:MetroWindow x:Class="Pixel.Automation.TestDataExplorer.Views.RenameDataSourceGroupView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:Controls="clr-namespace:MahApps.Metro.Controls;assembly=MahApps.Metro"
