+++ src/Pixel.Automation.TestDataExplorer.Views/TestDataExplorerHostView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="Pixel.Automation.TestData.Repository.Views.TestDataRepositoryHostView"
+﻿<UserControl x:Class="Pixel.Automation.TestDataExplorer.Views.TestDataExplorerHostView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
