+++ src/Pixel.Automation.TestDataExplorer.Views/TestDataExplorerHostView.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace Pixel.Automation.TestDataExplorer.Views
+{
+    /// <summary>
+    /// Interaction logic for TestDataExplorerHostView.xaml
+    /// </summary>
+    public partial class TestDataExplorerHostView : UserControl
+    {
+        public TestDataExplorerHostView()
+        {
+            InitializeComponent();
+        }     
+    }
+}
