+++ src/Pixel.Automation.TestDataExplorer.Views/TestDataModelEditorView.xaml
@@ -1,10 +1,10 @@
-﻿<UserControl x:Class="Pixel.Automation.TestData.Repository.Views.TestDataModelEditorView"
+﻿<UserControl x:Class="Pixel.Automation.TestDataExplorer.Views.TestDataModelEditorView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:cal="http://www.caliburnproject.org"
-             xmlns:local="clr-namespace:Pixel.Automation.TestData.Repository.Views"
+             xmlns:local="clr-namespace:Pixel.Automation.TestDataExplorer.Views"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
     <Grid>
