+++ src/Pixel.Automation.TestDataExplorer.Views/TestDataSourceBuilderView.xaml
@@ -1,9 +1,9 @@
-﻿<controls:MetroWindow x:Class="Pixel.Automation.TestData.Repository.Views.TestDataSourceBuilderView"
+﻿<controls:MetroWindow x:Class="Pixel.Automation.TestDataExplorer.Views.TestDataSourceBuilderView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:Pixel.Automation.TestData.Repository.Views"
+             xmlns:local="clr-namespace:Pixel.Automation.TestDataExplorer.Views"
              xmlns:cal="http://www.caliburnproject.org"
              xmlns:iconPacks="http://metro.mahapps.com/winfx/xaml/iconpacks" cal:Bind.AtDesignTime="True"
              xmlns:controls="clr-namespace:MahApps.Metro.Controls;assembly=MahApps.Metro"
