+++ src/Pixel.Automation.TestDataExplorer.Views/TestDataSourceBuilderView.xaml.cs
@@ -1,0 +1,13 @@
+﻿namespace Pixel.Automation.TestDataExplorer.Views
+{
+    /// <summary>
+    /// Interaction logic for TestDataSourceBuilderView.xaml
+    /// </summary>
+    public partial class TestDataSourceBuilderView
+    {
+        public TestDataSourceBuilderView()
+        {
+            InitializeComponent();
+        }
+    }
+}
