+++ src/Pixel.Automation.TestDataExplorer.Views/TestDataSourceView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="Pixel.Automation.TestData.Repository.Views.TestDataSourceView"
+﻿<UserControl x:Class="Pixel.Automation.TestDataExplorer.Views.TestDataSourceView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
