+++ src/Pixel.Automation.TestDataExplorer.Views/TestDataSourceView.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace Pixel.Automation.TestDataExplorer.Views
+{
+    /// <summary>
+    /// Interaction logic for TestDataSourceView.xaml
+    /// </summary>
+    public partial class TestDataSourceView : UserControl
+    {
+        public TestDataSourceView()
+        {
+            InitializeComponent();
+        }
+    }
+}
